+++ variants/forms.py
@@ -1098,16 +1098,16 @@ class SmallVariantPrioritizerFormMixin:
 
         self.fields["prio_hpo_terms"] = forms.CharField(
             label="HPO Terms",
-            help_text="HPO Term list.",
+            help_text="Enter a  list of HPO terms. The suggestions are based on HPO id and name, but only names are displayed. E.g., typing <span class='badge badge-secondary'>HP:0000001</span> will suggest <span class='badge badge-secondary'>All</span>, as will also typing <span class='badge badge-secondary'>all</span>.",
             widget=forms.Textarea(
                 attrs={
                     "placeholder": "Enter list of HPO terms here",
                     "rows": 3,
-                    "class": "form-control",
+                    "class": "form-control typeahead",
                 }
             ),
-            required=False,
             max_length=5000,
+            required=False,
         )
 
         self.fields["patho_enabled"] = forms.BooleanField(